requirements.txt
openai>=1.0.0
olmocr>=0.1.0
Pillow>=10.0.0
tqdm>=4.60.0
pypdfium2>=4.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
tiktoken>=0.5.0
gradio>=5.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
fastmcp>=3.0.0
sentence-transformers>=3.0.0
chromadb>=0.4.0
crawl4ai>=0.8.0
requests>=2.31.0
google-search-results>=2.4.0
slack-bolt>=1.18.0
aiohttp>=3.9.0
